# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GrabGo API
  description: |
    Backend API for GrabGo food delivery platform.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## User Roles
    - **customer**: Regular users who order food
    - **restaurant**: Restaurant owners/managers
    - **rider**: Delivery riders
    - **admin**: Platform administrators
  version: 1.0.0
  contact:
    name: GrabGo Support
  license:
    name: ISC
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.grabgo.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/register endpoints
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "+233201234567"
        isPhoneVerified:
          type: boolean
        isEmailVerified:
          type: boolean
        DateOfBirth:
          type: string
          format: date
        profilePicture:
          type: string
          format: uri
        isAdmin:
          type: boolean
        role:
          type: string
          enum: [customer, restaurant, rider, admin]
        isActive:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    Restaurant:
      type: object
      properties:
        _id:
          type: string
        restaurant_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        logo:
          type: string
          format: uri
        food_type:
          type: string
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        average_delivery_time:
          type: string
        delivery_fee:
          type: number
        min_order:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        rating:
          type: number
        is_open:
          type: boolean
        total_reviews:
          type: integer
...
    Order:
      type: object
      properties:
        _id:
          type: string
        orderNumber:
          type: string
        customer:
          $ref: '#/components/schemas/User'
        restaurant:
          $ref: '#/components/schemas/Restaurant'
        rider:
          $ref: '#/components/schemas/User'
        items:
          type: array
          items:
            type: object
            properties:
              food:
                type: string
              name:
                type: string
              quantity:
                type: integer
              price:
                type: number
              image:
                type: string
        subtotal:
          type: number
        deliveryFee:
          type: number
        tax:
          type: number
        totalAmount:
          type: number
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
          enum: [cash, card, mobile_money, mtn_momo, online]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, picked_up, on_the_way, delivered, cancelled]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: Get orders for current user
      description: Returns orders based on user role (customer, restaurant, or rider)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'

    post:
      tags: [Orders]
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurant, items, deliveryAddress, paymentMethod]
              properties:
                restaurant:
                  type: string
                  description: Restaurant ID
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [food, quantity]
                    properties:
                      food:
                        type: string
                        description: Food ID
                      quantity:
                        type: integer
                        minimum: 1
                deliveryAddress:
                  type: string
                paymentMethod:
                  type: string
                  enum: [cash, card, mobile_money, mtn_momo, online]
                notes:
                  type: string
                orderNumber:
                  type: string
      responses:
        '201':
          description: Order created successfully
        '400':
          description: Validation error
        '404':
          description: Restaurant or food not found
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order retrieved successfully
        '403':
          description: Not authorized
        '404':
          description: Order not found
  /orders/{orderId}/status:
    put:
      tags: [Orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, preparing, ready, picked_up, on_the_way, delivered, cancelled]
                cancellationReason:
                  type: string
      responses:
        '200':
          description: Status updated successfully
        '403':
          description: Customers can only cancel orders
        '404':
          description: Order not found
  /orders/{orderId}/assign-rider:
    put:
      tags: [Orders]
      summary: Assign rider to order (Admin/Rider only)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [riderId]
              properties:
                riderId:
                  type: string
      responses:
        '200':
          description: Rider assigned successfully
        '400':
          description: Invalid rider
        '404':
          description: Order not found

  # ============ PAYMENTS ============
...